+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 07-03-2024
+
+### Added
+- Add keymap `<leader>gb` to `:CurLnBlameToggle`.
+
 ## 06-27-2024
 
 ### Changed